Open up the self-documenting potential of bcrypt!
I love writing documentation, as I find it a great way to find bugs and design issues.  Design driven by API documentation FTW!

In this case, I had previously considered this design for the bcrypt integration, but rejected it because I didn't want to get too involved in modifying bcrypt itself.  However, writing the API documentation for the bcrypt binding clarified where that line was, which made me re-evaluate this decision once I started writing the API docs for the G3P.

What changed my mind is that even if the binding enforced a salt length of 72 bytes, one could still implement the standard bcrypt integrations on top of it.

My bright red line is whenever I introduce a change to the bcrypt binding that would prevent standard bcrypt integrations from being implemented.  My dull red line is that I'd really prefer that those integrations deal strictly with data encoding issues, and don't have to directly deal with cryptography.

The last thing I currently see like this is that "OrpheanBeholderScryDoubt" could be parameterized, but I don't see the the payoff.  Also since Electronic Codebook is used, care must be taken in selecting this string... another reason not to parameterize it.  Alternatively it might be safe to elide the output stage altogether, but this still feels like much too aggressive a change for no clear payoff.

Next up I will rework the G3Pb1 alfa message padding, as now it's out of sync with the ideal lengths for the bcrypt tags.