Padding Reform
I realized that one of the few SHA256 blocks that wasn't covered by any
self-documenting constant was the password itself.  This implied a very weak
form of multiparty computation where Alice has the deployment info and Bob
has the password, and they want to compute the hash together without directly
revealing the information to the other.

Here, Alice would hash the seguid and username, and provide this to Bob. Then
Bob would use this to hash the password, which would be returned to Alice for
key stretching.  Of course this also does reveal a hashed version of the
private information without any key-stretching, which may present a problem
if the information is potentially guessable.

It's not clear that this would actually be of any practical interest, let
alone a problem, but given the G3P's design philosophy, it seems suspect
enough to be worth eliminating.  In the new design, Bob will have to be
provided the first 32 bytes of the domain tag.

To do so, the SHA256 buffer-sync padding has been adapted to provide short
tags regularly throughout parameter transitions.  Incidentally, this hides
the length of the username in the case that it is being hidden from a
password cracker.

This totally breaks PHKDF's test suite, but these padding changes are
probably worth backporting.