Shorten test vector output
Comparing two blocks should be plenty, as the PHKDF key-stretching portion
much more comprehensively tests this core construction.

Not to mention, for generating extremely long inputs for more comprehensive
test vectors, I intend on modifying the test harness to be able to refer to
 the output stream of previous test cases, thus we can more comprehensively
test the expansion phase as well.