Yet Another Padding Reform
The echo-tags parameter had been bugging me for some time.  It's
functionality seemed rather dubious as it served largely the same purpose as
the roles vector. On the other hand I had never come up with a change that
seemed worth it.

Varying either the role (or the echo-tags) parameters would require a minimum
of seven SHA-256 blocks to be processed before the first output block is
obtained.  This is still true of the role vector, but now the echo-tag
parameter can be varied, with output coming in as little as two SHA-256
blocks, assuming the echo tag is 19 bytes or less.

To make this work well, the echoTag parameter replaces the domain tag
parameter for final output expansion, as needed to avoid the echo-args
gotcha.

For a bit of extra domain seperation on the final output, the first ~ 27
bytes of the original domain tag is included in the initial HMAC input
message. Truncating the domain tag avoids introducing any timing side
channels.

Also, if somebody is varying the echo tag, and the echo args
are too long, then some initial portion of the echo args can be replace
by a hash, and thus doesn't imply a complete cryptoacoustic repetition
of the domain tag every time the echo tag changes.