Strengthen the protection on the first 32 bytes of the p-box.
Always encoding the key on the left side of the p box after the first Blowfish
expansion phase seems likely better against hostile tag inputs. Keep in mind
that this well outside the primary intended use case, but there seems to
be little to no downside risk to this change and a potentially a lot of
upside benefit.