Add bcrypt-credentials parameter
Increment the number of parameters from a score to a full 21-gun salute. I have to make sure I
properly earn all those accusations of being guilty of kitchen-sink cryptographic engineering. ;-)

Incidentally also fixes a rather oof-y bug in which bcryptXsFree was never migrated to use
the new end-of-message padding.  I didn't bother trying to split this into a nicer sequence of
diffs.