Improve documentation and debugging tools
Debugging niceities like this are often so very worthwhile.  I probably should have taken
some of Gershom's wisdom on this whole project and not implemented the first version as
a fully fused implementation.

This is also probably an instance of me circumventing Dan's wisdom to write programs in the
most natural way possible first.  Though in my defense, accessing the right bits and
pieces is quite a bit more work here than in typical PLT scenarios.