WIP design document for G3P version 2
Yeah I know not all bcrypt parameters are correctly represented in the
overview, but I also don't want to overcomplicate it.  Perhaps I should
have developed this document in parallel with Version 2, but on the other
hand it didn't seem to fit the moment either.

While I might yet try to tweak the overview for more complete coverage,
everything needs to be exhaustively and correctly documented in later sections.