Add test for arbitary bitstrings
I generated the result myself, and though I'm fairly confident the answer
is correct, it's possible there is a bug lurking in here.  (Thus the desire
to verify this implementation against the official NIST test vectors)

On the other hand, this test does ensure that this library will at least
produce consistent results going forward.